Move IDataStore
